refactor: session table fetcher
