+++ packages/front-end/app/(sidebar-pages)/sessions/participant/page.tsx
@@ -4,7 +4,7 @@
 import { ErrorBoundary } from "react-error-boundary";
 import { Heading } from "@/components/Text";
 import { useQueryErrorResetBoundary } from "@tanstack/react-query";
-import SessionTableFetcher from "./_fetcher/SessionTableFetcher";
+import ParticipantSessionTableFetcher from "./_fetcher/ParticipantSessionTableFetcher";
 import Divider from "@/components/Divider";
 
 export default function Page() {
@@ -15,7 +15,7 @@ export default function Page() {
       <Heading>내가 참가한 세션</Heading>
       <Divider/>
       <ErrorBoundary fallback={<h1>에러</h1>} onReset={reset}>
-        <SessionTableFetcher />
+        <ParticipantSessionTableFetcher />
       </ErrorBoundary>
     </>
   );
