--- packages/front-end/util/getRoleFromURL.tsx
@@ -1,3 +1,0 @@
-export default function getRoleFromURL(pathname: string): string {
-  return pathname.split("/").at(-1) || "participant";
-}
